+++ mne/commands/mne_coreg.py
@@ -47,11 +47,16 @@ def run():
     parser.add_option('--trans', dest='trans', default=None,
                       help='Head<->MRI transform FIF file ("-trans.fif")')
     parser.add_option('--project-eeg', dest='project_eeg',
-                      action='store_true', default=False,
+                      action='store_true', default=None,
                       help="Project EEG electrodes to the head surface")
-    parser.add_option('--orient-points',
-                      action='store_true', default=False, dest='orient',
+    parser.add_option('--orient-to-surface',
+                      action='store_true', default=None,
+                      dest='orient_to_surface',
                       help='Orient points to the surface.')
+    parser.add_option('--scale-by-distance',
+                      action='store_true', default=None,
+                      dest='scale_by_distance',
+                      help='Scale points by distance from the surface.')
     parser.add_option('--verbose', action='store_true', dest='verbose',
                       help='Turn on verbose mode.')
 
@@ -80,8 +85,10 @@ def run():
             subjects_dir=subjects_dir,
             guess_mri_subject=options.guess_mri_subject,
             head_opacity=options.head_opacity, head_high_res=head_high_res,
-            trans=trans, scrollable=True, orient_points=options.orient,
-            project_eeg=options.project_eeg, verbose=options.verbose)
+            trans=trans, scrollable=True, project_eeg=options.project_eeg,
+            orient_to_surface=options.orient_to_surface,
+            scale_by_distance=options.scale_by_distance,
+            verbose=options.verbose)
     if is_main:
         sys.exit(0)